Learned about InitAbilityActorInfo and how it should be called depending on the owner of the Ability System
If the owner is Player State then Init Ability Actor Info is called in PossessedBy for server and OnRep_PlayerState for clients.

If the owner is Player Controlled Pawn then Init Ability Actor Info is called in PossessedBy for the server and AcknowledgePossesion for the client.

If the owner is the AI Controlled Pawn itself, it can be called in BeginPlay function.