Learned about Gameplay Effects with duration, periodic, stacking and infinite versions.
Duration without period is a gameplay effect that effects the target for a certain duration then removes the effect. Duration with period is that effects target for certain duration in certain periods. Stacking, just as the name describes, stacks effects. Infinite is also just as the name describes, effects infinitely unless removed manually.